dom/tv/test/mochitest/test_tv_get_channels.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test GetChannels for TV API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript" src="./head.js"></script>
<script type="application/javascript">

function runTest() {

  ok('tv' in navigator, "navigator.tv should exist.");

  navigator.tv.getTuners().then(
    function(aTuners) {
      ok(aTuners.length > 0, "Got at least 1 tuner.");

      aTuners[0].getSources().then(
        function(aSources) {
          ok(aSources.length > 0, "Got at least 1 source.");

          aSources[0].getChannels().then(
            function(aChannels) {
              ok(aChannels.length > 0, "Got at least 1 channel.");

              for (var i = 0; i < aChannels.length; i++) {
                var channel = aChannels[i];
                ok(channel instanceof TVChannel, "Channel " + i + " should be in right type.")
                ok('source' in channel, "Channel " + i + " should have a source.");
                ok('networkId' in channel, "Channel " + i + " should have a network ID.");
                ok('transportStreamId' in channel, "Channel " + i + " should have a transport stream ID.");
                ok('serviceId' in channel, "Channel " + i + " should have a service ID.");
                ok('type' in channel, "Channel " + i + " should have a type.");
                ok('name' in channel, "Channel " + i + " should have a name.");
                ok('number' in channel, "Channel " + i + " should have a number.");
              }

              SimpleTest.finish();
            },
            function(aError) {
              ok(false, "Error occurred when getting channels: " + aError);
              SimpleTest.finish();
            }
          );
        },
        function(aError) {
          ok(false, "Error occurred when getting sources: " + aError);
          SimpleTest.finish();
        }
      );
    },
    function(aError) {
      ok(false, "Error occurred when getting tuners: " + aError);
      SimpleTest.finish();
    }
  );
}

SimpleTest.waitForExplicitFinish();
prepareTest(runTest);

</script>
</pre>
</body>
</html>
